Rename metaphor.manual to metaphor.custom
