+++ metaphor/custom/governance/config.py
@@ -1,0 +1,52 @@
+from dataclasses import field as dataclass_field
+from typing import List
+
+from pydantic.dataclasses import dataclass
+
+from metaphor.common.base_extractor import BaseConfig
+from metaphor.common.models import DeserializableDatasetLogicalID
+
+
+@dataclass
+class Ownership:
+    # Type of ownership to assign
+    type: str
+
+    # Owner's email address
+    email: str
+
+
+@dataclass
+class ColumnTags:
+    # Name of the column
+    column: str
+
+    # List of associated tags
+    tags: List[str]
+
+
+@dataclass
+class Description:
+    # The description to assign
+    description: str
+
+    # The author's email address
+    email: str
+
+
+@dataclass
+class DatasetGovernance:
+    id: DeserializableDatasetLogicalID
+
+    ownerships: List[Ownership] = dataclass_field(default_factory=lambda: [])
+
+    tags: List[str] = dataclass_field(default_factory=lambda: [])
+
+    column_tags: List[ColumnTags] = dataclass_field(default_factory=lambda: [])
+
+    descriptions: List[Description] = dataclass_field(default_factory=lambda: [])
+
+
+@dataclass
+class CustomGovernanceConfig(BaseConfig):
+    datasets: List[DatasetGovernance]
