--- tests/manual/governance/test_extractor.py
@@ -1,18 +1,0 @@
-import pytest
-
-from metaphor.common.event_util import EventUtil
-from metaphor.manual.governance.config import ManualGovernanceConfig
-from metaphor.manual.governance.extractor import ManualGovernanceExtractor
-from tests.test_utils import load_json
-
-
-@pytest.mark.asyncio
-async def test_extractor(test_root_dir):
-    config = ManualGovernanceConfig.from_yaml_file(
-        f"{test_root_dir}/manual/governance/config.yml"
-    )
-    extractor = ManualGovernanceExtractor(config)
-
-    events = [EventUtil.trim_event(e) for e in await extractor.extract()]
-
-    assert events == load_json(f"{test_root_dir}/manual/governance/expected.json")
