--- tests/manual/lineage/expected.json
@@ -1,21 +1,0 @@
-[
-  {
-    "logicalId": {
-      "name": "foo",
-      "platform": "BIGQUERY"
-    },
-    "upstream": {
-      "sourceDatasets": [
-        "DATASET~856FE6D6B85A5758921110D2A1FC5E60",
-        "DATASET~2C268725B528CB0203E0F1AA22B95D65"
-      ]
-    }
-  },
-  {
-    "logicalId": {
-      "name": "bar",
-      "account": "account",
-      "platform": "SNOWFLAKE"
-    }
-  }
-]
